cleanup, rename & docs
