+++ scopes/ui-foundation/use-box/menu/index.ts
@@ -1,8 +1,5 @@
 export { Menu } from './menu';
-export type { MenuProps } from './menu';
-
-export { ConsumeMethodsMenu } from './consume-methods-menu';
-export type { ConsumeMethod } from './consume-methods-menu';
+export type { MenuProps, ConsumeMethod } from './menu';
 
 export { Import } from './import';
 export type { ImportProps } from './import';
