+++ scopes/ui-foundation/use-box/menu/menu.compositions.tsx
@@ -1,26 +1,31 @@
 import React from 'react';
 import { Menu } from './menu';
+import { Install } from './install';
+import { Import } from './import';
+
+const methods = [
+  {
+    Title: <img style={{ width: '30px' }} src="http://static.bit.dev/brands/logo-npm-new.svg" />,
+    Component: (
+      <Install
+        config={`npm config set @teambit:registry' https://node.bit.dev`}
+        componentName="radio"
+        packageManager="npm"
+        copyString="npm i @teambit/design.ui.radio"
+        registryName="@teambit"
+      />
+    ),
+    order: 1,
+  },
+  {
+    Title: <img style={{ width: '20px' }} src="https://static.bit.dev/brands/bit-logo-text.svg" />,
+    Component: (
+      <Import componentId="teambit.design/ui/radio" packageName="@teambit/design.ui.radio" componentName="radio" />
+    ),
+    order: 0,
+  },
+];
 
 export const MenuExample = () => {
-  return (
-    <Menu
-      style={{ width: 400 }}
-      packageName="@teambit/design.ui.input.radio"
-      componentId="teambit.design/ui/input/radio"
-      registryName="@teambit"
-      componentName="radio"
-    />
-  );
-};
-
-export const LongTooltipTextExample = () => {
-  return (
-    <Menu
-      style={{ width: 400 }}
-      packageName="@teambit/design.ui.input.radio.design.ui.input.radio.design.ui.input.radio.design.ui.input.radio.design.ui.input.radio.design.ui.input.radio"
-      componentId="teambit.design/ui/input/radio/design/ui/input/radio/design/ui/input/radio/design/ui/input/radio/design/ui/input/radio/design/ui/input/radio"
-      registryName="@teambit"
-      componentName="radio"
-    />
-  );
+  return <Menu style={{ width: 400 }} methods={methods} componentName="radio" />;
 };
