+++ scopes/ui-foundation/use-box/tab-content/index.ts
@@ -1,5 +1,5 @@
 export { TabContent } from './tab-content';
 export type { TabContentProps } from './tab-content';
 
-export { ConsumeMethodTemplate } from './consume-method-template';
-export type { ConsumeMethodTemplateProps } from './consume-method-template';
+export { ExpandableTabContent } from './expandable-tab-content';
+export type { ExpandableTabContentProps } from './expandable-tab-content';
