+++ scopes/ui-foundation/use-box/tab-content/tab-content.docs.mdx
@@ -4,6 +4,20 @@ labels: ['react', 'ui-component', 'component', 'tab']
 ---
 
 import { TabContent } from './tab-content';
+import { ExpandableTabContent } from './expandable-tab-content';
+
+The `ExpandableTabContent` is used when you want a bottom drawer that can be collapsed and expanded to show more info.
+
+```js live
+<ExpandableTabContent
+  style={{ width: 300 }}
+  content={<div>content of the tab</div>}
+  drawerTitle={<div>click here for more content</div>}
+  drawerContent={<div>some more content</div>}
+/>
+```
+
+The `TabContent` is used for the basic top-bottom layout for a tab in the menu.
 
 ```js live
 <TabContent bottom={<div>bottom</div>}>
